provide properties to allow overriding cart total
A quick fix to address an issue resulting from product default price being sent via cart total. Specifically for the payment-method update form within members areas, in cases where they user is offered the option to renew a stopped subscription. In such cases the price shown was being the product default as transmitted to the Vue components via the cart json passed from the backend. But if their existing subscription price was lower, that was not accurately reflected, so they would be reluctant to add their new payment method to restart their subscription because it appeared they would lose their low price. It turns out the renewal that did occur was actually properlly getting the price for their new subscription from their old one, as the OrderFormPayment component was not actually placing an order in such cases but rather calling a renew method. The OrderFormPayment component did not have handling imperfect cases which makes sense as this use of it is somewhat hacky. This commit enables resolving the issue by providing an option to pass a string with the price to display instead of the cart total. Hacky but it a proper solution would require more Vue work than I could do right now.